requirements.txt
acme==0.28.0
alembic==1.10.2
amplitude-analytics==1.0.0
amqp==2.6.1
anyio==3.6.2
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
blinker==1.3
boto3==1.16.56
botocore==1.19.63
Brlapi==0.8.3
cachetools==4.2.1
celery==4.4.7
certbot==0.28.0
certbot-nginx==0.28.0
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==1.2.3
configobj==5.0.6
crcmod==1.7
cryptography==3.2.1
cupshelpers==1.0
datadog-api-client==1.2.0
dbus-python==1.2.18
ddtrace==0.46.0
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.1.3
django-allow-cidr==0.3.1
django-autoslug==1.9.8
django-cacheops==5.1
django-cors-headers==3.5.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.21
django-extensions==3.1.1
django-mailjet==0.3.1
django-nine==0.2.4
django-redis==4.12.1
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.12.2
docutils==0.15.2
duplicity==0.8.21
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
entrypoints==0.4
executing==0.8.3
facebook-sdk==3.1.0
fastapi==0.94.0
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.6.0
flower==0.9.7
funcy==1.15
future==0.18.2
geographiclib==1.50
geopy==2.0.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
googlemaps==4.4.2
greenlet==2.0.2
gunicorn==20.0.4
gyp==0.1
h11==0.13.0
httplib2==0.19.0
humanize==4.1.0
idna==2.10
importlib-metadata==4.6.4
intervaltree==3.1.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
josepy==1.5.0
jsonpatch==1.10
jsonpointer==1.10
jsonschema==4.6.0
keyring==23.5.0
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
mailjet-rest==1.3.4
Mako==1.1.3
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
midtransclient==1.1.1
mistune==0.8.4
mock==4.0.2
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.0.22
mysqlclient==2.0.1
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
netifaces==0.11.0
notebook==6.4.12
numpy==1.22.4
oauthlib==3.2.0
olefile==0.46
packaging==20.9
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.9.3
parsedatetime==2.6
parso==0.8.3
pdfkit==0.6.1
pexpect==4.8.0
phonenumbers==8.12.17
pickleshare==0.7.5
Pillow==8.4.0
pipenv==11.9.0
platformdirs==2.5.1
prometheus-client==0.8.0
prompt-toolkit==3.0.29
protobuf==3.14.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.20
pycups==2.0.1
pydantic==1.10.6
Pygments==2.12.0
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==19.1.0
pyotp==2.6.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.15.0
python-json-logger==2.0.2
pytz==2020.4
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
redis==3.5.3
reportlab==3.6.8
requests==2.25.0
requests-toolbelt==0.9.1
rsa==3.4.2
s3transfer==0.3.7
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.5.post1
sqlparse==0.4.1
ssh-import-id==5.11
stack-data==0.3.0
starlette==0.26.0.post1
systemd-python==234
tenacity==7.0.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
twilio==6.51.1
typing_extensions==4.5.0
ua-parser==0.10.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.11
usb-creator==0.3.7
user-agents==2.2.0
uvicorn==0.15.0
vine==1.3.0
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wrapt==1.14.1
wsproto==1.0.0
xdg==5
xid==1.0.4
-e git+https://github.com/dwin/xid_py@239a096de0ee6c8308773e5c8e5c6f06d72c67dc#egg=xid_py
xkit==0.0.0
zipp==1.0.0
zope.component==4.6.2
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.2.0
zope.proxy==4.3.5
